requirements.txt
pandas
ta
joblib
python-binance
websocket-client

